refactor: split CollaborationHub and UserAchievements into sub-components
CollaborationHub (828 -> ~220 lines):
- Extracted OpportunitiesSection, CollaborationsSection, NetworkingSection,
  CreateRequestForm, and OpportunityDetailModal into
  src/components/collaboration/

UserAchievements (806 -> ~200 lines):
- Extracted XPProgressWheel, AnalyticsMetrics, OnboardingChecklist,
  MilestoneBadges, and ShareBadgeModal into
  src/components/user/achievements/

EventCreation was already refactored into src/components/common/EventCreation/
with 5+ sub-components in a prior pass.

All existing functionality preserved. Build should pass with no errors.

Fixes #4783
