--- .github/workflows/auto-resolve-conflicts.yml
@@ -1,33 +1,0 @@
-name: Auto-Resolve Merge Conflicts
-on:
-  workflow_dispatch:
-  schedule:
-    - cron: '0 * * * *'
-permissions:
-  contents: write
-  pull-requests: write
-jobs:
-  resolve:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          token: ${{ secrets.GH_TOKEN }}
-      - name: Install Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: npm
-      - name: Auto resolve all open PRs
-        env:
-          GH_TOKEN: ${{ secrets.GH_TOKEN }}
-        run: |
-          gh pr list --repo ${{ github.repository }} --state open --base master --json number,headRefName | jq -c '.[]' | while read -r pr; do
-            num=$(echo $pr | jq -r '.number')
-            branch=$(echo $pr | jq -r '.headRefName')
-            gh pr checkout $num
-            if git merge origin/master --no-commit --no-ff 2>&1 | grep -q "CONFLICT"; then
-              for file in $(git diff --name-only --diff-filter=U); do
-                # Auto resolve by keeping current branch version
-                git checkout --ours "$file"
-                git add "$file"
