--- .github/workflows/gssoc-approved-label.yml
@@ -1,24 +1,0 @@
-name: Add GSSoC Approved Label
-
-on:
-  pull_request:
-    types:
-      - closed
-
-permissions:
-  pull-requests: write
-  issues: write
-
-jobs:
-  add-label:
-    if: github.event.pull_request.merged == true
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Add gssoc:approved label
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          gh pr edit ${{ github.event.pull_request.number }} \
-            --repo ${{ github.repository }} \
-            --add-label "gssoc:approved"
