--- .github/workflows/markdown-linter.yml
@@ -1,24 +1,0 @@
-name: Markdown Linting
-
-on:
-  push:
-    branches: [ master ]
-    paths:
-      - '**/*.md'
-  pull_request:
-    branches: [ master ]
-    paths:
-      - '**/*.md'
-
-jobs:
-  markdown-lint:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout Code
-        uses: actions/checkout@v4
-
-      - name: Run Markdown Lint
-        uses: DavidAnson/markdownlint-cli2-action@v16
-        with:
-          globs: '**/*.md'
-          config: '.github/.markdownlint.json'
