--- .github/workflows/release.yml
@@ -1,37 +1,0 @@
-name: Release
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  release-build:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: "22.x"
-          cache: npm
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Build project
-        run: npm run build
-        env:
-          NODE_ENV: production
-          JWT_SECRET: "mock-jwt-secret-for-ci-build-verification-only"
-
-      - name: Archive build directory
-        run: tar -czf build.tar.gz build
-
-      - name: Upload Release Asset
-        uses: softprops/action-gh-release@v2
-        with:
-          files: build.tar.gz
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
