--- .github/workflows/welcome-new-issues.yml
@@ -1,34 +1,0 @@
-name: Welcome new issues
-
-on:
-  issues:
-    types:
-      - opened
-
-permissions:
-  issues: write
-  contents: read
-
-jobs:
-  comment-on-issue:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Post Discord invite comment
-        uses: actions/github-script@v7
-        with:
-          script: |
-            const body = [
-              'Thanks for opening this issue.',
-              '',
-              'If you want to connect with the Eventra community or follow updates, please join our Discord server:',
-              'https://discord.gg/6MQ9r5nHT',
-              '',
-              'We will review your issue as soon as possible.'
-            ].join('\n');
-
-            await github.rest.issues.createComment({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              issue_number: context.issue.number,
-              body,
-            });
